create useActivityInfo hook
