+++ src/app/active-mode/Views/CurrentActivity.tsx
@@ -1,18 +1,16 @@
 import { DestinyAccount } from 'app/accounts/destiny-account';
+import { useActivityInfo } from 'app/active-mode/Views/activity-util';
 import ActivityInformation from 'app/active-mode/Views/current-activity/ActivityInformation';
 import VendorBounties from 'app/active-mode/Views/current-activity/VendorBounties';
 import styles from 'app/active-mode/Views/CurrentActivity.m.scss';
-import { getCurrentActivity } from 'app/bungie-api/destiny2-api';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import BungieImage from 'app/dim-ui/BungieImage';
 import CollapsibleTitle from 'app/dim-ui/CollapsibleTitle';
 import { t } from 'app/i18next-t';
 import { InventoryBuckets } from 'app/inventory/inventory-buckets';
 import { DimStore } from 'app/inventory/store-types';
-import { refresh } from 'app/shell/refresh';
 import { RootState } from 'app/store/types';
-import { DestinyCharacterActivitiesComponent } from 'bungie-api-ts/destiny2';
-import React, { useEffect, useRef, useState } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
 
 interface ProvidedProps {
@@ -33,58 +31,8 @@ function mapStateToProps(state: RootState): StoreProps {
 
 type Props = ProvidedProps & StoreProps;
 
-const ACTIVITY_REFRESH_RATE = 10_000;
-let cooldown = Date.now();
-let intervalId = 0;
-let firstLoad = true;
-
-async function refreshActivity({ account, store }: { account: DestinyAccount; store: DimStore }) {
-  const profileInfo = await getCurrentActivity(account);
-  return profileInfo.characterActivities.data?.[store.id];
-}
-
 function CurrentActivity({ account, store, defs, buckets }: Props) {
-  const [activityInfo, setActivityInfo] = useState<
-    DestinyCharacterActivitiesComponent | undefined
-  >();
-  const prevActivityInfo = useRef<DestinyCharacterActivitiesComponent | undefined>();
-
-  useEffect(() => {
-    if (firstLoad) {
-      refreshActivity({ account, store }).then(setActivityInfo);
-    }
-
-    window.clearInterval(intervalId);
-    intervalId = window.setInterval(() => {
-      // If the activity mode just changed, trust the new one and don't ping for a little.
-      // Sometimes there is a weird cache issue, where even after some time it flips back and forth with the last
-      if (Date.now() - cooldown < ACTIVITY_REFRESH_RATE * 6) {
-        return;
-      }
-
-      refreshActivity({ account, store }).then(setActivityInfo);
-    }, ACTIVITY_REFRESH_RATE);
-
-    return () => {
-      window.clearInterval(intervalId);
-      firstLoad = true;
-    };
-  }, [account, store]);
-
-  useEffect(() => {
-    if (!activityInfo) {
-      return;
-    }
-    if (
-      !firstLoad &&
-      prevActivityInfo.current?.currentActivityHash !== activityInfo.currentActivityHash
-    ) {
-      refresh();
-      cooldown = Date.now();
-    }
-    firstLoad = false;
-    prevActivityInfo.current = activityInfo;
-  }, [activityInfo]);
+  const activityInfo = useActivityInfo({ account, store });
 
   if (!defs) {
     return null;
