+++ src/app/active-mode/Views/activity-util.ts
@@ -1,14 +1,74 @@
+import { DestinyAccount } from 'app/accounts/destiny-account';
 import { currentAccountSelector } from 'app/accounts/selectors';
+import { getCurrentActivity } from 'app/bungie-api/destiny2-api';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { bucketsSelector } from 'app/inventory/selectors';
 import { DimStore } from 'app/inventory/store-types';
+import { refresh } from 'app/shell/refresh';
 import { RootState } from 'app/store/types';
 import { toVendor } from 'app/vendors/d2-vendors';
 import { VendorItem } from 'app/vendors/vendor-item';
 import { DestinyCharacterActivitiesComponent } from 'bungie-api-ts/destiny2';
 import { ItemCategoryHashes } from 'data/d2/generated-enums';
 import pursuitsInfo from 'data/d2/pursuits.json';
+import { useEffect, useRef, useState } from 'react';
 import { createSelector } from 'reselect';
+
+const ACTIVITY_REFRESH_RATE = 10_000;
+let cooldown = Date.now();
+let intervalId = 0;
+let firstLoad = true;
+
+async function refreshActivity({ account, store }: { account: DestinyAccount; store: DimStore }) {
+  const profileInfo = await getCurrentActivity(account);
+  return profileInfo.characterActivities.data?.[store.id];
+}
+
+export function useActivityInfo({ account, store }: { account: DestinyAccount; store: DimStore }) {
+  const [activityInfo, setActivityInfo] = useState<
+    DestinyCharacterActivitiesComponent | undefined
+  >();
+  const prevActivityInfo = useRef<DestinyCharacterActivitiesComponent | undefined>();
+
+  useEffect(() => {
+    if (firstLoad) {
+      refreshActivity({ account, store }).then(setActivityInfo);
+    }
+
+    window.clearInterval(intervalId);
+    intervalId = window.setInterval(() => {
+      // If the activity mode just changed, trust the new one and don't ping for a little.
+      // Sometimes there is a weird cache issue, where even after some time it flips back and forth with the last
+      if (Date.now() - cooldown < ACTIVITY_REFRESH_RATE * 6) {
+        return;
+      }
+
+      refreshActivity({ account, store }).then(setActivityInfo);
+    }, ACTIVITY_REFRESH_RATE);
+
+    return () => {
+      window.clearInterval(intervalId);
+      firstLoad = true;
+    };
+  }, [account, store]);
+
+  useEffect(() => {
+    if (!activityInfo) {
+      return;
+    }
+    if (
+      !firstLoad &&
+      prevActivityInfo.current?.currentActivityHash !== activityInfo.currentActivityHash
+    ) {
+      refresh();
+      cooldown = Date.now();
+    }
+    firstLoad = false;
+    prevActivityInfo.current = activityInfo;
+  }, [activityInfo]);
+
+  return activityInfo;
+}
 
 export const purchasableBountiesSelector = (store: DimStore) =>
   createSelector(
